refactor: api 호출 비동기 처리
