+++ app/src/main/java/campus/tech/kakao/map/data/PlaceRepositoryImpl.kt
@@ -4,13 +4,22 @@ import android.content.ContentValues
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
+import androidx.lifecycle.viewModelScope
+import campus.tech.kakao.map.BuildConfig
+import campus.tech.kakao.map.PlaceApplication
+import campus.tech.kakao.map.data.net.KakaoApiClient
 import campus.tech.kakao.map.util.PlaceContract
 import campus.tech.kakao.map.domain.model.Place
 import campus.tech.kakao.map.domain.repository.PlaceRepository
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
 class PlaceRepositoryImpl(context: Context):
     SQLiteOpenHelper(context, PlaceContract.DATABASE_NAME, null, 1),
     PlaceRepository {
+
     override fun onCreate(db: SQLiteDatabase?) {
         db?.execSQL(PlaceContract.CREATE_QUERY)
         db?.execSQL(PlaceContract.CREATE_LOG_QUERY)
@@ -21,48 +30,26 @@ class PlaceRepositoryImpl(context: Context):
         db?.execSQL(PlaceContract.DROP_LOG_QUERY)
         onCreate(db)
     }
-
-    override fun getPlaces(placeName: String): List<Place> {
-        val places = mutableListOf<Place>()
-        val cursor = readableDatabase.query(
-            PlaceContract.TABLE_NAME,
-            null, "${PlaceContract.COLUMN_NAME} LIKE ?", arrayOf("${placeName}%"), null, null, null
-        )
-        cursor?.use {
-            while (it.moveToNext()) {
-                val id = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_ID))
-                val name = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_NAME))
-                val place = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_LOCATION))
-                val type = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_TYPE))
-                val xPos = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_X_POS))
-                val yPos = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_Y_POS))
-                places.add(Place(id, name, place, type, xPos,yPos))
+    override suspend fun getPlaces(keyword: String): List<Place> =
+        withContext(Dispatchers.IO){
+            val resultPlaces = mutableListOf<Place>()
+            for (page in 1..3) {
+                val response = KakaoApiClient.api.getSearchKeyword(
+                    key = BuildConfig.KAKAO_REST_API_KEY,
+                    query = keyword,
+                    size = 15,
+                    page = page
+                )
+                if (response.isSuccessful) {
+                    response.body()?.documents?.let { resultPlaces.addAll(it) }
+                } else throw RuntimeException("통신 에러 발생")
             }
+            updatePlaces(resultPlaces)
+            resultPlaces
         }
-        return places
-    }
 
-    override fun getAllPlaces(): List<Place> {
-        val places = mutableListOf<Place>()
-        val cursor = readableDatabase.query(
-            PlaceContract.TABLE_NAME,
-            null, null, null, null, null, null
-        )
-        cursor?.use {
-            while (it.moveToNext()) {
-                val id = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_ID))
-                val name = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_NAME))
-                val place = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_LOCATION))
-                val type = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_TYPE))
-                val xPos = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_X_POS))
-                val yPos = it.getString(it.getColumnIndexOrThrow(PlaceContract.COLUMN_Y_POS))
-                places.add(Place(id, name, place, type, xPos,yPos))
-            }
-        }
-        return places
-    }
 
-    override fun updatePlaces(places: List<Place>) {
+    override suspend fun updatePlaces(places: List<Place>) {
         val db = writableDatabase
 
         db.execSQL(PlaceContract.DELETE_QUERY)
