+++ app/src/main/java/campus/tech/kakao/map/domain/repository/PlaceRepository.kt
@@ -3,11 +3,10 @@ package campus.tech.kakao.map.domain.repository
 import campus.tech.kakao.map.domain.model.Place
 
 interface PlaceRepository {
-    fun getPlaces(placeName: String): List<Place>
-    fun getAllPlaces():List<Place>
-    fun updatePlaces(places:List<Place>)
+    suspend fun getPlaces(placeName: String): List<Place>
+    suspend fun updatePlaces(places:List<Place>)
+    fun getLogs(): List<Place>
     fun updateLogs(placeLog: List<Place>)
     fun removeLog(id: String)
-    fun getLogs(): List<Place>
 
 }
